+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- (cli) Improved warning messages printed when an optional background update of a local mirror of a Git repo fails (e.g. due to lack of internet access) to better reassure the user that it might not be a problem.
+
 ## 0.8.0-beta.0 - 2025-03-19
 
 ### Added